% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-frequency.R
\name{base_frequency}
\alias{base_frequency}
\title{base_frequency}
\usage{
base_frequency(fasfile)
}
\arguments{
\item{fasfile}{fasta file}
}
\value{
matrix
}
\description{
caculate base frequency
}
\author{
Xinzhi Mo
}
