% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gb2fasta.R
\name{download_genbank}
\alias{download_genbank}
\title{download_genebank}
\usage{
download_genbank(accession)
}
\arguments{
\item{accession}{accession number}
}
\description{
download genebank files
}
\author{
Xinzhi Mo
}
