% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gb2fasta.R
\name{gb_fasta}
\alias{gb_fasta}
\title{gb_fasta}
\usage{
gb_fasta(gbfile)
}
\arguments{
\item{gbfile}{genebank file}
}
\value{
fasta file
}
\description{
convert genebank files to fasta format
}
\author{
Xinzhi Mo
}
