% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gb2fasta.R
\name{read.fasta}
\alias{read.fasta}
\title{read.fasta}
\usage{
read.fasta(fasfile)
}
\arguments{
\item{fasfile}{fasta file}
}
\value{
sequence
}
\description{
extract the sequence in the fasta file
}
\author{
Xinzhi Mo
}
