% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq-alignment.R
\name{seq_align}
\alias{seq_align}
\title{seq_align}
\usage{
seq_align(V, W)
}
\arguments{
\item{V}{sequence1}

\item{W}{sequence2}
}
\value{
score matrix and align sequences
}
\description{
sequence alignment
}
\author{
Xinzhi Mo

Xinzhi Mo
}
